Refactor test to be reusable
